+++ src/LightEngine/Upgrade.h
@@ -5,7 +5,7 @@ class Upgrade : public Entity {
 
     int m_price;
     int m_type;
-    int m_value;
+    float m_value;
     int m_lane;
     std::string m_text;
     int m_MaxLevel;
@@ -18,7 +18,7 @@ public:
 
     void SetPrice(int value) { m_price = value; };
     void SetType(int value) { m_type = value; };
-    void SetValue(int value) { m_value = value; };
+    void SetValue(float value) { m_value = value; };
     void SetLane(int value);
     void SetText(std::string value) { m_text = value; };
     void SetMaxLevel(int value) { m_MaxLevel = value; };
